Move mutation to x.go
